[elasticsearch-utils] enable dynamic lookup of nodes
Currently, heroic is deployed internally with the list of seeds being all the data nodes.

During an incident when capacity was added to an already "red" cluster, puppet modified the heroic config
with the new data nodes, and restarted the api. The API was never able to be successfully started bringing down
everything.

This change will allow us to have the master nodes that rarely change become the seeds and the data nodes "sniffed" out.

https://www.elastic.co/guide/en/elasticsearch/client/java-api/current/transport-client.html